refactored for legibility, added healthchecks and custom conf
This adds "groups" of nodes so when acorn is scaled up and down
the additions and removals are deterministic.

This also adds healthchecks to the redis pods.

This adds the basic ablity to scale up a redis cluster adding both
leaders and replicas.

Added ability to pass in additional configuration for all cluster
roles.